sante_digitale/requirements.txt
# Core Django
Django==3.1.12
djangorestframework==3.12.4
django-cors-headers==3.10.0
django-environ==0.11.2

# MongoDB Integration
djongo==1.3.7
pymongo==4.8.0

# Authentication
PyJWT==2.8.0
djangorestframework-simplejwt==4.8.0

# API Documentation
drf-yasg==1.20.0
drf-spectacular==0.22.0

# Firebase/Notifications
firebase-admin==6.5.0
celery==5.4.0
redis==5.0.8

# ML/Data Processing
numpy==2.0.0
pandas==2.2.2
scikit-learn==1.5.1

# Development
whitenoise==6.7.0
gunicorn==22.0.0
python-dotenv==1.0.1

# Testing
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.0

# Security
cryptography==43.0.0
